refactor: split templates into multiple files
