--- crates/cdk-ldk-node/src/web/handlers.rs
@@ -1,2389 +1,0 @@
-use std::collections::HashMap;
-use std::str::FromStr;
-use std::sync::Arc;
-
-use axum::body::Body;
-use axum::extract::{Query, State};
-use axum::http::StatusCode;
-use axum::response::{Html, Response};
-use axum::Form;
-use ldk_node::bitcoin::secp256k1::PublicKey;
-use ldk_node::bitcoin::Address;
-use ldk_node::lightning::ln::msgs::SocketAddress;
-use ldk_node::lightning::offers::offer::Offer;
-use ldk_node::lightning_invoice::Bolt11Invoice;
-use ldk_node::payment::{PaymentDirection, PaymentKind, PaymentStatus};
-use ldk_node::UserChannelId;
-use maud::html;
-use serde::{Deserialize, Serialize};
-
-// Custom deserializer for optional u32 that handles empty strings
-fn deserialize_optional_u32<'de, D>(deserializer: D) -> Result<Option<u32>, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    let opt = Option::<String>::deserialize(deserializer)?;
-    match opt.as_deref() {
-        None | Some("") => Ok(None),
-        Some(s) => s.parse::<u32>().map(Some).map_err(serde::de::Error::custom),
-    }
-}
-
-// Custom deserializer for optional u64 that handles empty strings
-fn deserialize_optional_u64<'de, D>(deserializer: D) -> Result<Option<u64>, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    let opt = Option::<String>::deserialize(deserializer)?;
-    match opt.as_deref() {
-        None | Some("") => Ok(None),
-        Some(s) => s.parse::<u64>().map(Some).map_err(serde::de::Error::custom),
-    }
-}
-
-// Custom deserializer for optional f64 that handles empty strings
-fn deserialize_optional_f64<'de, D>(deserializer: D) -> Result<Option<f64>, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    let opt = Option::<String>::deserialize(deserializer)?;
-    match opt.as_deref() {
-        None | Some("") => Ok(None),
-        Some(s) => s.parse::<f64>().map(Some).map_err(serde::de::Error::custom),
-    }
-}
-
-use cdk_common::util::hex;
-
-use crate::web::templates::{
-    error_message, form_card, format_msats_as_btc, format_sats_as_btc, info_card, layout,
-    payment_list_item, success_message,
-};
-use crate::CdkLdkNode;
-
-#[derive(Clone)]
-pub struct AppState {
-    pub node: Arc<CdkLdkNode>,
-}
-
-#[derive(Debug)]
-pub struct UsageMetrics {
-    pub lightning_inflow_24h: u64,
-    pub lightning_outflow_24h: u64,
-    pub lightning_inflow_all_time: u64,
-    pub lightning_outflow_all_time: u64,
-    pub onchain_inflow_24h: u64,
-    pub onchain_outflow_24h: u64,
-    pub onchain_inflow_all_time: u64,
-    pub onchain_outflow_all_time: u64,
-}
-
-/// Get paginated payments with efficient filtering and sorting
-fn get_paginated_payments_streaming(
-    node: &ldk_node::Node,
-    filter: &str,
-    skip: usize,
-    take: usize,
-) -> (Vec<ldk_node::payment::PaymentDetails>, usize) {
-    // Create filter predicate - note LDK expects &&PaymentDetails
-    let filter_fn = match filter {
-        "incoming" => {
-            |p: &&ldk_node::payment::PaymentDetails| p.direction == PaymentDirection::Inbound
-        }
-        "outgoing" => {
-            |p: &&ldk_node::payment::PaymentDetails| p.direction == PaymentDirection::Outbound
-        }
-        _ => |_: &&ldk_node::payment::PaymentDetails| true,
-    };
-
-    // Get filtered payments from LDK
-    let filtered_payments = node.list_payments_with_filter(filter_fn);
-
-    // Create sorted index to avoid cloning payments during sort
-    let mut time_indexed: Vec<_> = filtered_payments
-        .iter()
-        .enumerate()
-        .map(|(idx, payment)| (payment.latest_update_timestamp, idx))
-        .collect();
-
-    // Sort by timestamp (newest first)
-    time_indexed.sort_unstable_by(|a, b| b.0.cmp(&a.0));
-
-    let total_count = time_indexed.len();
-
-    // Extract only the payments we need for this page
-    let page_payments: Vec<_> = time_indexed
-        .into_iter()
-        .skip(skip)
-        .take(take)
-        .map(|(_, idx)| filtered_payments[idx].clone())
-        .collect();
-
-    (page_payments, total_count)
-}
-
-/// Calculate usage metrics from payment history
-fn calculate_usage_metrics(payments: &[ldk_node::payment::PaymentDetails]) -> UsageMetrics {
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    let now = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs();
-    let twenty_four_hours_ago = now.saturating_sub(24 * 60 * 60);
-
-    let mut metrics = UsageMetrics {
-        lightning_inflow_24h: 0,
-        lightning_outflow_24h: 0,
-        lightning_inflow_all_time: 0,
-        lightning_outflow_all_time: 0,
-        onchain_inflow_24h: 0,
-        onchain_outflow_24h: 0,
-        onchain_inflow_all_time: 0,
-        onchain_outflow_all_time: 0,
-    };
-
-    for payment in payments {
-        // Only count successful payments
-        if payment.status != PaymentStatus::Succeeded {
-            continue;
-        }
-
-        let amount_sats = payment.amount_msat.unwrap_or(0) / 1000;
-        let is_recent = payment.latest_update_timestamp >= twenty_four_hours_ago;
-
-        match &payment.kind {
-            // Lightning payments (BOLT11, BOLT12, Spontaneous)
-            PaymentKind::Bolt11 { .. }
-            | PaymentKind::Bolt12Offer { .. }
-            | PaymentKind::Bolt12Refund { .. }
-            | PaymentKind::Spontaneous { .. }
-            | PaymentKind::Bolt11Jit { .. } => match payment.direction {
-                PaymentDirection::Inbound => {
-                    metrics.lightning_inflow_all_time += amount_sats;
-                    if is_recent {
-                        metrics.lightning_inflow_24h += amount_sats;
-                    }
-                }
-                PaymentDirection::Outbound => {
-                    metrics.lightning_outflow_all_time += amount_sats;
-                    if is_recent {
-                        metrics.lightning_outflow_24h += amount_sats;
-                    }
-                }
-            },
-            // On-chain payments
-            PaymentKind::Onchain { .. } => match payment.direction {
-                PaymentDirection::Inbound => {
-                    metrics.onchain_inflow_all_time += amount_sats;
-                    if is_recent {
-                        metrics.onchain_inflow_24h += amount_sats;
-                    }
-                }
-                PaymentDirection::Outbound => {
-                    metrics.onchain_outflow_all_time += amount_sats;
-                    if is_recent {
-                        metrics.onchain_outflow_24h += amount_sats;
-                    }
-                }
-            },
-        }
-    }
-
-    metrics
-}
-
-pub async fn dashboard(State(state): State<AppState>) -> Result<Html<String>, StatusCode> {
-    let node = &state.node.inner;
-
-    let _node_id = node.node_id().to_string();
-    let alias = node
-        .node_alias()
-        .map(|a| a.to_string())
-        .unwrap_or_else(|| "No alias set".to_string());
-
-    let listening_addresses: Vec<String> = state
-        .node
-        .inner
-        .announcement_addresses()
-        .as_ref()
-        .unwrap_or(&vec![])
-        .iter()
-        .map(|a| a.to_string())
-        .collect();
-
-    let (num_peers, num_connected_peers) =
-        node.list_peers()
-            .iter()
-            .fold((0, 0), |(mut peers, mut connected), p| {
-                if p.is_connected {
-                    connected += 1;
-                }
-                peers += 1;
-                (peers, connected)
-            });
-
-    let (num_active_channels, num_inactive_channels) =
-        node.list_channels()
-            .iter()
-            .fold((0, 0), |(mut active, mut inactive), c| {
-                if c.is_usable {
-                    active += 1;
-                } else {
-                    inactive += 1;
-                }
-                (active, inactive)
-            });
-
-    let balances = node.list_balances();
-
-    // Calculate payment metrics for dashboard
-    let all_payments = node.list_payments_with_filter(|_| true);
-    let metrics = calculate_usage_metrics(&all_payments);
-
-    let content = html! {
-        h2 style="text-align: center; margin-bottom: 3rem;" { "Dashboard" }
-
-        // Balance Summary as metric cards
-        div class="card" {
-            h2 { "Balance Summary" }
-            div class="metrics-container" {
-                div class="metric-card" {
-                    div class="metric-value" { (format_sats_as_btc(balances.total_lightning_balance_sats)) }
-                    div class="metric-label" { "Lightning Balance" }
-                }
-                div class="metric-card" {
-                    div class="metric-value" { (format_sats_as_btc(balances.total_onchain_balance_sats)) }
-                    div class="metric-label" { "On-chain Balance" }
-                }
-                div class="metric-card" {
-                    div class="metric-value" { (format_sats_as_btc(balances.spendable_onchain_balance_sats)) }
-                    div class="metric-label" { "Spendable Balance" }
-                }
-                div class="metric-card" {
-                    div class="metric-value" { (format_sats_as_btc(balances.total_lightning_balance_sats + balances.total_onchain_balance_sats)) }
-                    div class="metric-label" { "Combined Total" }
-                }
-            }
-        }
-
-        // Node Information - new layout based on Figma design
-        section class="node-info-section" {
-            div class="node-info-main-container" {
-                // Left side - Node avatar and info
-                div class="node-info-left" {
-                    div class="node-avatar" {
-                        img src="/static/images/nut.png" alt="Node Avatar" class="avatar-image";
-                    }
-                    div class="node-details" {
-                        h2 class="node-name" { (alias.clone()) }
-                        p class="node-address" {
-                            "Listening Address: "
-                            (listening_addresses.first().unwrap_or(&"127.0.0.1:8090".to_string()))
-                        }
-                    }
-                }
-
-                // Middle - Gray container with spinning globe animation
-                div class="node-content-box" {
-                    div class="globe-container" {
-                        svg aria-hidden="true" style="position: absolute; width: 0; height: 0; overflow: hidden;" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" {
-                            defs {
-                                symbol id="icon-world" viewBox="0 0 216 100" {
-                                    title { "world" }
-                                    g fill-rule="nonzero" {
-                                        path d="M48 94l-3-4-2-14c0-3-1-5-3-8-4-5-6-9-4-11l1-4 1-3c2-1 9 0 11 1l3 2 2 3 1 2 8 2c1 1 2 2 0 7-1 5-2 7-4 7l-2 3-2 4-2 3-2 1c-2 2-2 9 0 10v1l-3-2zM188 90l3-2h1l-4 2zM176 87h2l-1 1-1-1zM195 86l3-2-2 2h-1zM175 83l-1-2-2-1-6 1c-5 1-5 1-5-2l1-4 2-2 4-3c5-4 9-5 9-3 0 3 3 3 4 1s1-2 1 0l3 4c2 4 1 6-2 10-4 3-7 4-8 1zM100 80c-2-4-4-11-3-14l-1-6c-1-1-2-3-1-4 0-2-4-3-9-3-4 0-5 0-7-3-1-2-2-4-1-7l3-6 3-3c1-2 10-4 11-2l6 3 5-1c3 1 4 0 5-1s-1-2-2-2l-4-1c0-1 3-3 6-2 3 0 3 0 2-2-2-2-6-2-7 0l-2 2-1 2-3-2-3-3c-1 0-1 1 1 2l1 2-2-1c-4-3-6-2-8 1-2 2-4 3-5 1-1-1 0-4 2-4l2-2 1-2 3-2 3-2 2 1c3 0 7-3 5-4l-1-3h-1l-1 3-2 2h-1l-2-1c-2-1-2-1 1-4 5-4 6-4 11-3 4 1 4 1 2 2v1l3-1 6-1c5 0 6-1 5-2l2 1c1 2 2 2 2 1-2-4 12-7 14-4l11 1 29 3 1 2-3 3c-2 0-2 0-1 1l1 3h-2c-1-1-2-3-1-4h-4l-6 2c-1 1-1 1 2 2 3 2 4 6 1 8v3c1 3 0 3-3 0s-4-1-2 3c3 4 3 7-2 8-5 2-4 1-2 5 2 3 0 5-3 4l-2-1-2-2-1-1-1-1-2-2c-1-2-1-2-4 0-2 1-3 4-3 5-1 3-1 3-3 1l-2-4c0-2-1-3-2-3l-1-1-4-2-6-1-4-2c-1 1 3 4 5 4h2c1 1 0 2-1 4-3 2-7 4-8 3l-7-10 5 10c2 2 3 3 5 2 3 0 2 1-2 7-4 4-4 5-4 8 1 3 1 4-1 6l-2 3c0 2-6 9-8 9l-3-2zm22-51l-2-3-1-1v-1c-2 0-2 2-1 4 2 3 4 4 4 1z" {}
-                                        path d="M117 75c-1-2 0-6 2-7h2l-2 5c0 2-1 3-2 1zM186 64h-3c-2 0-6-3-5-5 1-1 6 1 7 3l2 3-2-1zM160 62h2c1 1 0 1-1 1l-1-1zM154 57l-1-2c2 2 3 1 2-2l-2-3 2 2 1 4 1 3v2l-3-4zM161 59c-1-1-1-2 1-4 3-3 4-3 4 0 0 4-2 6-5 4zM167 59l1-1 1 1-1 1-1-1zM176 59l1-1v2l-1-1zM141 52l1-1v2l-1-1zM170 52l1-1v2l-1-1zM32 50c-1-2-4-3-6-4-4-1-5-3-7-6l-3-5-2-2c-1-3-1-6 2-9 1-1 2-3 1-5 0-4-3-5-8-4H4l2-2 1-1 1-1 2-1c1-2 7-2 23-1 12 1 12 1 12-1h1c1 1 2 2 3 1l1 1-3 1c-2 0-8 4-8 5l2 1 2 3 4-3c3-4 4-4 5-3l3 1 1 2 1 2c3 0-1 2-4 2-2 0-2 0-2 2 1 1 0 2-2 2-4 1-12 9-12 12 0 2 0 2-1 1 0-2-2-3-6-2-3 0-4 1-4 3-2 4 0 6 3 4 3-1 3-1 2 1s-1 2 1 2l1 2 1 3 1 1-3-2zm8-24l1-1c0-1-4-3-5-2l1 1v2c-1 1-1 1 0 0h3zM167 47v-3l1 2c1 2 0 3-1 1z" {}
-                                        path d="M41 43h2l-1 1-1-1zM37 42v-1l2 1h-2zM16 38l1-1v2l-1-1zM172 32l2-3h1c1 2 0 4-3 4v-1zM173 26h2l-1 1-1-1zM56 22h2l-2 1v-1zM87 19l1-2 1 3-1 1-1-2zM85 19l1-1v1l-1 1v-1zM64 12l1-3c2 0-1-4-3-4s-2 0 0-1V3l-6 2c-3 1-3 1-2-1 2-1 4-2 15-2h14c0 2-6 7-10 9l-5 2-2 1-2-2zM53 12l1-1c2 0-1-3-3-3-2-1-1-1 1-1l4 2c2 1 2 1 1 3-2 1-4 2-4 0zM80 12l1-1 1 1-1 1-1-1zM36 8h-2V7c1-1 7 0 7 1h-5zM116 7l1-1v1l-1 1V7zM50 5h2l-1 1-1-1zM97 5l2-1c0-1 1-1 0 0l-2 1z" {}
-                                    }
-                                }
-                                symbol id="icon-repeated-world" viewBox="0 0 432 100" {
-                                    use href="#icon-world" x="0" {}
-                                    use href="#icon-world" x="189" {}
-                                }
-                            }
-                        }
-                        span class="world" {
-                            span class="images" {
-                                svg { use href="#icon-repeated-world" {} }
-                            }
-                        }
-                    }
-                }
-            }
-
-            // Right side - Connections metrics
-            aside class="node-metrics" {
-                div class="card" {
-                    h3 { "Connections" }
-                    div class="metrics-container" {
-                        div class="metric-card" {
-                            div class="metric-value" { (format!("{}/{}", num_connected_peers, num_peers)) }
-                            div class="metric-label" { "Connected Peers" }
-                        }
-                        div class="metric-card" {
-                            div class="metric-value" { (format!("{}/{}", num_active_channels, num_active_channels + num_inactive_channels)) }
-                            div class="metric-label" { "Active Channels" }
-                        }
-                    }
-                }
-            }
-        }
-
-        // Lightning Network Activity as metric cards
-        div class="card" {
-            h2 { "Lightning Network Activity" }
-            div class="metrics-container" {
-                div class="metric-card" {
-                    div class="metric-value" { (format_sats_as_btc(metrics.lightning_inflow_24h)) }
-                    div class="metric-label" { "24h LN Inflow" }
-                }
-                div class="metric-card" {
-                    div class="metric-value" { (format_sats_as_btc(metrics.lightning_outflow_24h)) }
-                    div class="metric-label" { "24h LN Outflow" }
-                }
-                div class="metric-card" {
-                    div class="metric-value" { (format_sats_as_btc(metrics.lightning_inflow_all_time)) }
-                    div class="metric-label" { "All-time LN Inflow" }
-                }
-                div class="metric-card" {
-                    div class="metric-value" { (format_sats_as_btc(metrics.lightning_outflow_all_time)) }
-                    div class="metric-label" { "All-time LN Outflow" }
-                }
-            }
-        }
-
-        // On-chain Activity as metric cards
-        div class="card" {
-            h2 { "On-chain Activity" }
-            div class="metrics-container" {
-                div class="metric-card" {
-                    div class="metric-value" { (format_sats_as_btc(metrics.onchain_inflow_24h)) }
-                    div class="metric-label" { "24h On-chain Inflow" }
-                }
-                div class="metric-card" {
-                    div class="metric-value" { (format_sats_as_btc(metrics.onchain_outflow_24h)) }
-                    div class="metric-label" { "24h On-chain Outflow" }
-                }
-                div class="metric-card" {
-                    div class="metric-value" { (format_sats_as_btc(metrics.onchain_inflow_all_time)) }
-                    div class="metric-label" { "All-time On-chain Inflow" }
-                }
-                div class="metric-card" {
-                    div class="metric-value" { (format_sats_as_btc(metrics.onchain_outflow_all_time)) }
-                    div class="metric-label" { "All-time On-chain Outflow" }
-                }
-            }
-
-        }
-    };
-
-    Ok(Html(layout("Dashboard", content).into_string()))
-}
-
-pub async fn get_new_address(State(state): State<AppState>) -> Result<Html<String>, StatusCode> {
-    let address_result = state.node.inner.onchain_payment().new_address();
-
-    let content = match address_result {
-        Ok(address) => {
-            html! {
-                (success_message(&format!("New address generated: {address}")))
-                div class="card" {
-                    h2 { "Bitcoin Address" }
-                    div class="info-item" {
-                        span class="info-label" { "Address:" }
-                        span class="info-value" style="font-family: monospace; font-size: 0.9rem;" { (address.to_string()) }
-                    }
-                }
-                div class="card" {
-                    a href="/onchain" { button { "← Back to On-chain" } }
-                    " "
-                    a href="/onchain/new-address" { button { "Generate Another Address" } }
-                }
-            }
-        }
-        Err(e) => {
-            html! {
-                (error_message(&format!("Failed to generate address: {e}")))
-                div class="card" {
-                    a href="/onchain" { button { "← Back to On-chain" } }
-                }
-            }
-        }
-    };
-
-    Ok(Html(layout("New Address", content).into_string()))
-}
-
-pub async fn balance_page(State(state): State<AppState>) -> Result<Html<String>, StatusCode> {
-    let balances = state.node.inner.list_balances();
-    let channels = state.node.inner.list_channels();
-
-    let (num_active_channels, num_inactive_channels) =
-        channels
-            .iter()
-            .fold((0, 0), |(mut active, mut inactive), c| {
-                if c.is_usable {
-                    active += 1;
-                } else {
-                    inactive += 1;
-                }
-                (active, inactive)
-            });
-
-    let content = if channels.is_empty() {
-        html! {
-            h2 style="text-align: center; margin-bottom: 3rem;" { "Lightning" }
-
-            // Quick Actions section - matching dashboard style
-            div class="card" style="margin-bottom: 2rem;" {
-                h2 { "Quick Actions" }
-                div style="display: flex; gap: 1rem; margin-top: 1rem; flex-wrap: wrap;" {
-                    a href="/channels/open" style="text-decoration: none; flex: 1; min-width: 200px;" {
-                        button class="button-primary" style="width: 100%;" { "Open Channel" }
-                    }
-                    a href="/invoices" style="text-decoration: none; flex: 1; min-width: 200px;" {
-                        button class="button-primary" style="width: 100%;" { "Create Invoice" }
-                    }
-                    a href="/payments/send" style="text-decoration: none; flex: 1; min-width: 200px;" {
-                        button class="button-primary" style="width: 100%;" { "Make Lightning Payment" }
-                    }
-                }
-            }
-
-            // Balance Information as metric cards
-            div class="card" {
-                h2 { "Balance Information" }
-                div class="metrics-container" {
-                    div class="metric-card" {
-                        div class="metric-value" { (format_sats_as_btc(balances.total_lightning_balance_sats)) }
-                        div class="metric-label" { "Lightning Balance" }
-                    }
-                    div class="metric-card" {
-                        div class="metric-value" { (format!("{}", num_active_channels + num_inactive_channels)) }
-                        div class="metric-label" { "Total Channels" }
-                    }
-                    div class="metric-card" {
-                        div class="metric-value" { (format!("{}", num_active_channels)) }
-                        div class="metric-label" { "Active Channels" }
-                    }
-                    div class="metric-card" {
-                        div class="metric-value" { (format!("{}", num_inactive_channels)) }
-                        div class="metric-label" { "Inactive Channels" }
-                    }
-                }
-            }
-
-            div class="card" {
-                p { "No channels found. Create your first channel to start using Lightning Network." }
-            }
-        }
-    } else {
-        html! {
-            h2 style="text-align: center; margin-bottom: 3rem;" { "Lightning" }
-
-            // Quick Actions section - matching dashboard style
-            div class="card" style="margin-bottom: 2rem;" {
-                h2 { "Quick Actions" }
-                div style="display: flex; gap: 1rem; margin-top: 1rem; flex-wrap: wrap;" {
-                    a href="/channels/open" style="text-decoration: none; flex: 1; min-width: 200px;" {
-                        button class="button-primary" style="width: 100%;" { "Open Channel" }
-                    }
-                    a href="/invoices" style="text-decoration: none; flex: 1; min-width: 200px;" {
-                        button class="button-primary" style="width: 100%;" { "Create Invoice" }
-                    }
-                    a href="/payments/send" style="text-decoration: none; flex: 1; min-width: 200px;" {
-                        button class="button-primary" style="width: 100%;" { "Make Lightning Payment" }
-                    }
-                }
-            }
-
-            // Balance Information as metric cards
-            div class="card" {
-                h2 { "Balance Information" }
-                div class="metrics-container" {
-                    div class="metric-card" {
-                        div class="metric-value" { (format_sats_as_btc(balances.total_lightning_balance_sats)) }
-                        div class="metric-label" { "Lightning Balance" }
-                    }
-                    div class="metric-card" {
-                        div class="metric-value" { (format!("{}", num_active_channels + num_inactive_channels)) }
-                        div class="metric-label" { "Total Channels" }
-                    }
-                    div class="metric-card" {
-                        div class="metric-value" { (format!("{}", num_active_channels)) }
-                        div class="metric-label" { "Active Channels" }
-                    }
-                    div class="metric-card" {
-                        div class="metric-value" { (format!("{}", num_inactive_channels)) }
-                        div class="metric-label" { "Inactive Channels" }
-                    }
-                }
-            }
-
-            div class="card" {
-                h2 { "Channel Details" }
-
-                // Channels list
-                @for channel in &channels {
-                    div class="channel-item" {
-                        div class="channel-header" {
-                            span class="channel-id" { "Channel ID: " (channel.channel_id.to_string()) }
-                            @if channel.is_usable {
-                                span class="status-badge status-active" { "Active" }
-                            } @else {
-                                span class="status-badge status-inactive" { "Inactive" }
-                            }
-                        }
-                        div class="info-item" {
-                            span class="info-label" { "Counterparty" }
-                            span class="info-value" style="font-family: monospace; font-size: 0.85rem;" { (channel.counterparty_node_id.to_string()) }
-                        }
-                        @if let Some(short_channel_id) = channel.short_channel_id {
-                            div class="info-item" {
-                                span class="info-label" { "Short Channel ID" }
-                                span class="info-value" { (short_channel_id.to_string()) }
-                            }
-                        }
-                        div class="balance-info" {
-                            div class="balance-item" {
-                                div class="balance-amount" { (format_sats_as_btc(channel.outbound_capacity_msat / 1000)) }
-                                div class="balance-label" { "Outbound" }
-                            }
-                            div class="balance-item" {
-                                div class="balance-amount" { (format_sats_as_btc(channel.inbound_capacity_msat / 1000)) }
-                                div class="balance-label" { "Inbound" }
-                            }
-                            div class="balance-item" {
-                                div class="balance-amount" { (format_sats_as_btc(channel.channel_value_sats)) }
-                                div class="balance-label" { "Total" }
-                            }
-                        }
-                        @if channel.is_usable {
-                            div style="margin-top: 1rem; display: flex; gap: 0.5rem;" {
-                                a href=(format!("/channels/close?channel_id={}&node_id={}", channel.user_channel_id.0, channel.counterparty_node_id)) {
-                                    button style="background: #dc3545;" { "Close Channel" }
-                                }
-                                a href=(format!("/channels/force-close?channel_id={}&node_id={}", channel.user_channel_id.0, channel.counterparty_node_id)) {
-                                    button style="background: #d63384;" title="Force close should not be used if normal close is preferred. Force close will broadcast the latest commitment transaction immediately." { "Force Close" }
-                                }
-                            }
-                        }
-                    }
-                }
-
-            }
-        }
-    };
-
-    Ok(Html(layout("Lightning", content).into_string()))
-}
-
-#[derive(Deserialize, Serialize)]
-pub struct SendOnchainActionForm {
-    address: String,
-    #[serde(deserialize_with = "deserialize_optional_u64")]
-    amount_sat: Option<u64>,
-    send_action: String,
-}
-
-pub async fn onchain_page(
-    State(state): State<AppState>,
-    query: Query<HashMap<String, String>>,
-) -> Result<Html<String>, StatusCode> {
-    let balances = state.node.inner.list_balances();
-    let action = query
-        .get("action")
-        .map(|s| s.as_str())
-        .unwrap_or("overview");
-
-    let mut content = html! {
-        h2 style="text-align: center; margin-bottom: 3rem;" { "On-chain" }
-
-        // Quick Actions section - matching dashboard style
-        div class="card" style="margin-bottom: 2rem;" {
-            h2 { "Quick Actions" }
-            div style="display: flex; gap: 1rem; margin-top: 1rem; flex-wrap: wrap;" {
-                a href="/onchain?action=receive" style="text-decoration: none; flex: 1; min-width: 200px;" {
-                    button class="button-primary" style="width: 100%;" { "Receive Bitcoin" }
-                }
-                a href="/onchain?action=send" style="text-decoration: none; flex: 1; min-width: 200px;" {
-                    button class="button-primary" style="width: 100%;" { "Send Bitcoin" }
-                }
-            }
-        }
-
-        // On-chain Balance as metric cards
-        div class="card" {
-            h2 { "On-chain Balance" }
-            div class="metrics-container" {
-                div class="metric-card" {
-                    div class="metric-value" { (format_sats_as_btc(balances.total_onchain_balance_sats)) }
-                    div class="metric-label" { "Total Balance" }
-                }
-                div class="metric-card" {
-                    div class="metric-value" { (format_sats_as_btc(balances.spendable_onchain_balance_sats)) }
-                    div class="metric-label" { "Spendable Balance" }
-                }
-            }
-        }
-    };
-
-    match action {
-        "send" => {
-            // Show send form above balance section
-            content = html! {
-                h2 style="text-align: center; margin-bottom: 3rem;" { "On-chain" }
-
-                // Quick Actions section - matching dashboard style
-                div class="card" style="margin-bottom: 2rem;" {
-                    h2 { "Quick Actions" }
-                    div style="display: flex; gap: 1rem; margin-top: 1rem; flex-wrap: wrap;" {
-                        a href="/onchain?action=receive" style="text-decoration: none; flex: 1; min-width: 200px;" {
-                            button class="button-primary" style="width: 100%;" { "Receive Bitcoin" }
-                        }
-                        a href="/onchain?action=send" style="text-decoration: none; flex: 1; min-width: 200px;" {
-                            button class="button-primary" style="width: 100%;" { "Send Bitcoin" }
-                        }
-                    }
-                }
-
-                // Send form above balance
-                (form_card(
-                    "Send On-chain Payment",
-                    html! {
-                        form method="post" action="/onchain/send" {
-                            div class="form-group" {
-                                label for="address" { "Recipient Address" }
-                                input type="text" id="address" name="address" required placeholder="bc1..." {}
-                            }
-                            div class="form-group" {
-                                label for="amount_sat" { "Amount (sats)" }
-                                input type="number" id="amount_sat" name="amount_sat" placeholder="0" {}
-                            }
-                            input type="hidden" id="send_action" name="send_action" value="send" {}
-                            div style="display: flex; justify-content: space-between; gap: 1rem; margin-top: 2rem;" {
-                                a href="/onchain" { button type="button" { "Cancel" } }
-                                div style="display: flex; gap: 0.5rem;" {
-                                    button type="submit" onclick="document.getElementById('send_action').value='send'" { "Send Payment" }
-                                    button type="submit" onclick="document.getElementById('send_action').value='send_all'; document.getElementById('amount_sat').value=''" { "Send All" }
-                                }
-                            }
-                        }
-                    }
-                ))
-
-                // On-chain Balance as metric cards
-                div class="card" {
-                    h2 { "On-chain Balance" }
-                    div class="metrics-container" {
-                        div class="metric-card" {
-                            div class="metric-value" { (format_sats_as_btc(balances.total_onchain_balance_sats)) }
-                            div class="metric-label" { "Total Balance" }
-                        }
-                        div class="metric-card" {
-                            div class="metric-value" { (format_sats_as_btc(balances.spendable_onchain_balance_sats)) }
-                            div class="metric-label" { "Spendable Balance" }
-                        }
-                    }
-                }
-            };
-        }
-        "receive" => {
-            // Show generate address form above balance section
-            content = html! {
-                h2 style="text-align: center; margin-bottom: 3rem;" { "On-chain" }
-
-                // Quick Actions section - matching dashboard style
-                div class="card" style="margin-bottom: 2rem;" {
-                    h2 { "Quick Actions" }
-                    div style="display: flex; gap: 1rem; margin-top: 1rem; flex-wrap: wrap;" {
-                        a href="/onchain?action=receive" style="text-decoration: none; flex: 1; min-width: 200px;" {
-                            button class="button-primary" style="width: 100%;" { "Receive Bitcoin" }
-                        }
-                        a href="/onchain?action=send" style="text-decoration: none; flex: 1; min-width: 200px;" {
-                            button class="button-primary" style="width: 100%;" { "Send Bitcoin" }
-                        }
-                    }
-                }
-
-                // Generate address form above balance
-                (form_card(
-                    "Generate New Address",
-                    html! {
-                        form method="post" action="/onchain/new-address" {
-                            p style="margin-bottom: 2rem;" { "Click the button below to generate a new Bitcoin address for receiving on-chain payments." }
-                            div style="display: flex; justify-content: space-between; gap: 1rem;" {
-                                a href="/onchain" { button type="button" { "Cancel" } }
-                                button class="button-primary" type="submit" { "Generate New Address" }
-                            }
-                        }
-                    }
-                ))
-
-                // On-chain Balance as metric cards
-                div class="card" {
-                    h2 { "On-chain Balance" }
-                    div class="metrics-container" {
-                        div class="metric-card" {
-                            div class="metric-value" { (format_sats_as_btc(balances.total_onchain_balance_sats)) }
-                            div class="metric-label" { "Total Balance" }
-                        }
-                        div class="metric-card" {
-                            div class="metric-value" { (format_sats_as_btc(balances.spendable_onchain_balance_sats)) }
-                            div class="metric-label" { "Spendable Balance" }
-                        }
-                    }
-                }
-            };
-        }
-        _ => {
-            // Show overview with just the balance and quick actions at the top
-            // No additional content needed since quick actions are now at the top
-        }
-    }
-
-    Ok(Html(layout("On-chain", content).into_string()))
-}
-
-pub async fn post_send_onchain(
-    State(_state): State<AppState>,
-    Form(form): Form<SendOnchainActionForm>,
-) -> Result<Response, StatusCode> {
-    // Show confirmation screen first, then redirect to the confirmation handler
-    let encoded_form =
-        serde_urlencoded::to_string(&form).map_err(|_| StatusCode::INTERNAL_SERVER_ERROR)?;
-
-    Ok(Response::builder()
-        .status(StatusCode::FOUND)
-        .header("Location", format!("/onchain/confirm?{}", encoded_form))
-        .body(Body::empty())
-        .unwrap())
-}
-
-#[derive(Deserialize)]
-pub struct ConfirmOnchainForm {
-    address: String,
-    // #[serde(deserialize_with = "deserialize_optional_u64")]
-    amount_sat: Option<u64>,
-    send_action: String,
-    confirmed: Option<String>,
-}
-
-pub async fn onchain_confirm_page(
-    State(state): State<AppState>,
-    query: Query<ConfirmOnchainForm>,
-) -> Result<Response, StatusCode> {
-    let form = query.0;
-
-    // If user confirmed, execute the transaction
-    if form.confirmed.as_deref() == Some("true") {
-        return execute_onchain_transaction(State(state), form).await;
-    }
-
-    // Validate address
-    let _address = match Address::from_str(&form.address) {
-        Ok(addr) => addr,
-        Err(e) => {
-            let content = html! {
-                (error_message(&format!("Invalid address: {e}")))
-                div class="card" {
-                    a href="/onchain?action=send" { button { "← Back" } }
-                }
-            };
-            return Ok(Response::builder()
-                .status(StatusCode::BAD_REQUEST)
-                .header("content-type", "text/html")
-                .body(Body::from(
-                    layout("Send On-chain Error", content).into_string(),
-                ))
-                .unwrap());
-        }
-    };
-
-    let balances = state.node.inner.list_balances();
-    let spendable_balance = balances.spendable_onchain_balance_sats;
-
-    // Calculate transaction details
-    let (amount_to_send, is_send_all) = if form.send_action == "send_all" {
-        (spendable_balance, true)
-    } else {
-        let amount = form.amount_sat.unwrap_or(0);
-        if amount > spendable_balance {
-            let content = html! {
-                (error_message(&format!("Insufficient funds. Requested: {}, Available: {}",
-                    format_sats_as_btc(amount), format_sats_as_btc(spendable_balance))))
-                div class="card" {
-                    a href="/onchain?action=send" { button { "← Back" } }
-                }
-            };
-            return Ok(Response::builder()
-                .status(StatusCode::BAD_REQUEST)
-                .header("content-type", "text/html")
-                .body(Body::from(
-                    layout("Send On-chain Error", content).into_string(),
-                ))
-                .unwrap());
-        }
-        (amount, false)
-    };
-
-    let confirmation_url = if form.send_action == "send_all" {
-        format!(
-            "/onchain/confirm?address={}&send_action={}&confirmed=true",
-            urlencoding::encode(&form.address),
-            form.send_action
-        )
-    } else {
-        format!(
-            "/onchain/confirm?address={}&amount_sat={}&send_action={}&confirmed=true",
-            urlencoding::encode(&form.address),
-            form.amount_sat.unwrap_or(0),
-            form.send_action
-        )
-    };
-
-    let content = html! {
-        h2 style="text-align: center; margin-bottom: 3rem;" { "Confirm On-chain Transaction" }
-
-        div class="card" style="border: 2px solid hsl(var(--primary)); background-color: hsl(var(--primary) / 0.05);" {
-            h2 { "⚠️ Transaction Confirmation" }
-            p style="color: hsl(var(--muted-foreground)); margin-bottom: 1.5rem;" {
-                "Please review the transaction details carefully before proceeding. This action cannot be undone."
-            }
-        }
-
-        (info_card(
-            "Transaction Details",
-            vec![
-                ("Recipient Address", form.address.clone()),
-                ("Amount to Send", if is_send_all {
-                    format!("{} (All available funds)", format_sats_as_btc(amount_to_send))
-                } else {
-                    format_sats_as_btc(amount_to_send)
-                }),
-                ("Current Spendable Balance", format_sats_as_btc(spendable_balance)),
-            ]
-        ))
-
-        @if is_send_all {
-            div class="card" style="border: 1px solid hsl(32.6 75.4% 55.1%); background-color: hsl(32.6 75.4% 55.1% / 0.1);" {
-                h3 style="color: hsl(32.6 75.4% 55.1%);" { "Send All Notice" }
-                p style="color: hsl(32.6 75.4% 55.1%);" {
-                    "This transaction will send all available funds to the recipient address. "
-                    "Network fees will be deducted from the total amount automatically."
-                }
-            }
-        }
-
-        div class="card" {
-            div style="display: flex; justify-content: space-between; gap: 1rem; margin-top: 2rem;" {
-                a href="/onchain?action=send" {
-                    button type="button" class="button-secondary" { "← Cancel" }
-                }
-                div style="display: flex; gap: 0.5rem;" {
-                    a href=(confirmation_url) {
-                        button class="button-primary" {
-                            "✓ Confirm & Send Transaction"
-                        }
-                    }
-                }
-            }
-        }
-    };
-
-    Ok(Response::builder()
-        .header("content-type", "text/html")
-        .body(Body::from(
-            layout("Confirm Transaction", content).into_string(),
-        ))
-        .unwrap())
-}
-
-async fn execute_onchain_transaction(
-    State(state): State<AppState>,
-    form: ConfirmOnchainForm,
-) -> Result<Response, StatusCode> {
-    tracing::info!(
-        "Web interface: Executing on-chain transaction to address={}, send_action={}, amount_sat={:?}",
-        form.address,
-        form.send_action,
-        form.amount_sat
-    );
-
-    let address = match Address::from_str(&form.address) {
-        Ok(addr) => addr,
-        Err(e) => {
-            tracing::warn!(
-                "Web interface: Invalid address for on-chain transaction: {}",
-                e
-            );
-            let content = html! {
-                (error_message(&format!("Invalid address: {e}")))
-                div class="card" {
-                    a href="/onchain" { button { "← Back" } }
-                }
-            };
-            return Ok(Response::builder()
-                .status(StatusCode::BAD_REQUEST)
-                .header("content-type", "text/html")
-                .body(Body::from(
-                    layout("Send On-chain Error", content).into_string(),
-                ))
-                .unwrap());
-        }
-    };
-
-    // Handle send all action
-    let txid_result = if form.send_action == "send_all" {
-        tracing::info!(
-            "Web interface: Sending all available funds to {}",
-            form.address
-        );
-        // Use send_all_to_address function
-        state.node.inner.onchain_payment().send_all_to_address(
-            address.assume_checked_ref(),
-            false,
-            None,
-        )
-    } else {
-        let amount_sats = form.amount_sat.ok_or(StatusCode::BAD_REQUEST)?;
-        tracing::info!(
-            "Web interface: Sending {} sats to {}",
-            amount_sats,
-            form.address
-        );
-        // Use regular send_to_address function
-        state.node.inner.onchain_payment().send_to_address(
-            address.assume_checked_ref(),
-            amount_sats,
-            None,
-        )
-    };
-
-    let content = match txid_result {
-        Ok(txid) => {
-            if form.send_action == "send_all" {
-                tracing::info!(
-                    "Web interface: Successfully sent all available funds, txid={}",
-                    txid
-                );
-            } else {
-                tracing::info!(
-                    "Web interface: Successfully sent {} sats, txid={}",
-                    form.amount_sat.unwrap_or(0),
-                    txid
-                );
-            }
-            let amount = form.amount_sat;
-            html! {
-                        (success_message("Transaction sent successfully!"))
-                        (info_card(
-                            "Transaction Details",
-                            vec![
-                                ("Transaction ID", txid.to_string()),
-                                ("Amount", if form.send_action == "send_all" {
-                                    format!("{} (All available funds)", format_msats_as_btc(amount.unwrap_or(0)))
-                                } else {
-                                    format_sats_as_btc(form.amount_sat.unwrap_or(0))
-                                }),
-                                ("Recipient", form.address),
-                            ]
-                        ))
-                        div class="card" {
-                            a href="/onchain" { button { "← Back to On-chain" } }
-                        }
-            }
-        }
-        Err(e) => {
-            tracing::error!("Web interface: Failed to send on-chain transaction: {}", e);
-            html! {
-                (error_message(&format!("Failed to send payment: {e}")))
-                div class="card" {
-                    a href="/onchain" { button { "← Try Again" } }
-                }
-            }
-        }
-    };
-
-    Ok(Response::builder()
-        .header("content-type", "text/html")
-        .body(Body::from(
-            layout("Send On-chain Result", content).into_string(),
-        ))
-        .unwrap())
-}
-
-pub async fn channels_page(State(_state): State<AppState>) -> Result<Response, StatusCode> {
-    // Redirect to the balance page since channels are now part of the Lightning section
-    Ok(Response::builder()
-        .status(StatusCode::FOUND)
-        .header("Location", "/balance")
-        .body(Body::empty())
-        .unwrap())
-}
-
-pub async fn open_channel_page(State(_state): State<AppState>) -> Result<Html<String>, StatusCode> {
-    let content = form_card(
-        "Open New Channel",
-        html! {
-            form method="post" action="/channels/open" {
-                div class="form-group" {
-                    label for="node_id" { "Node Public Key" }
-                    input type="text" id="node_id" name="node_id" required placeholder="02..." {}
-                }
-                div class="form-group" {
-                    label for="address" { "Node Address" }
-                    input type="text" id="address" name="address" required placeholder="127.0.0.1" {}
-                }
-                div class="form-group" {
-                    label for="port" { "Port" }
-                    input type="number" id="port" name="port" required value="9735" {}
-                }
-                div class="form-group" {
-                    label for="amount_btc" { "Channel Size" }
-                    input type="number" id="amount_sats" name="amount_sats" required placeholder="₿0" step="1" {}
-                }
-                div class="form-group" {
-                    label for="push_btc" { "Push Amount (optional)" }
-                    input type="number" id="push_btc" name="push_btc" placeholder="₿0" step="1" {}
-                }
-                button type="submit" { "Open Channel" }
-                " "
-                a href="/balance" { button type="button" { "Cancel" } }
-            }
-        },
-    );
-
-    Ok(Html(layout("Open Channel", content).into_string()))
-}
-
-#[derive(Deserialize)]
-pub struct OpenChannelForm {
-    node_id: String,
-    address: String,
-    port: u32,
-    amount_sats: u64,
-    #[serde(deserialize_with = "deserialize_optional_u64")]
-    push_btc: Option<u64>,
-}
-
-pub async fn post_open_channel(
-    State(state): State<AppState>,
-    Form(form): Form<OpenChannelForm>,
-) -> Result<Response, StatusCode> {
-    tracing::info!(
-        "Web interface: Attempting to open channel to node_id={}, address={}:{}, amount_sats={}, push_btc={:?}",
-        form.node_id,
-        form.address,
-        form.port,
-        form.amount_sats,
-        form.push_btc
-    );
-
-    let pubkey = match PublicKey::from_str(&form.node_id) {
-        Ok(pk) => pk,
-        Err(e) => {
-            tracing::warn!("Web interface: Invalid node public key provided: {}", e);
-            let content = html! {
-                (error_message(&format!("Invalid node public key: {e}")))
-                div class="card" {
-                    a href="/channels/open" { button { "← Try Again" } }
-                }
-            };
-            return Ok(Response::builder()
-                .status(StatusCode::BAD_REQUEST)
-                .header("content-type", "text/html")
-                .body(Body::from(
-                    layout("Open Channel Error", content).into_string(),
-                ))
-                .unwrap());
-        }
-    };
-
-    let socket_addr = match SocketAddress::from_str(&format!("{}:{}", form.address, form.port)) {
-        Ok(addr) => addr,
-        Err(e) => {
-            tracing::warn!("Web interface: Invalid address:port combination: {}", e);
-            let content = html! {
-                (error_message(&format!("Invalid address:port combination: {e}")))
-                div class="card" {
-                    a href="/channels/open" { button { "← Try Again" } }
-                }
-            };
-            return Ok(Response::builder()
-                .status(StatusCode::BAD_REQUEST)
-                .header("content-type", "text/html")
-                .body(Body::from(
-                    layout("Open Channel Error", content).into_string(),
-                ))
-                .unwrap());
-        }
-    };
-
-    // First connect to the peer
-    tracing::info!(
-        "Web interface: Connecting to peer {} at {}",
-        pubkey,
-        socket_addr
-    );
-    if let Err(e) = state.node.inner.connect(pubkey, socket_addr.clone(), true) {
-        tracing::error!("Web interface: Failed to connect to peer {}: {}", pubkey, e);
-        let content = html! {
-            (error_message(&format!("Failed to connect to peer: {e}")))
-            div class="card" {
-                a href="/channels/open" { button { "← Try Again" } }
-            }
-        };
-        return Ok(Response::builder()
-            .status(StatusCode::INTERNAL_SERVER_ERROR)
-            .header("content-type", "text/html")
-            .body(Body::from(
-                layout("Open Channel Error", content).into_string(),
-            ))
-            .unwrap());
-    }
-
-    // Convert Bitcoin to millisatoshis (1 BTC = 100,000,000,000 msats)
-
-    // Then open the channel
-    tracing::info!(
-        "Web interface: Opening announced channel to {} with amount {} sats and push amount {:?} msats",
-        pubkey,
-        form.amount_sats,
-        form.push_btc.map(|a| a * 1000)
-    );
-    let channel_result = state.node.inner.open_announced_channel(
-        pubkey,
-        socket_addr,
-        form.amount_sats,
-        form.push_btc.map(|a| a * 1000), // Pass None when empty, Some(value) when provided
-        None,
-    );
-
-    let content = match channel_result {
-        Ok(user_channel_id) => {
-            tracing::info!(
-                "Web interface: Successfully initiated channel opening with user_channel_id={} to {}",
-                user_channel_id.0,
-                pubkey
-            );
-            html! {
-                (success_message("Channel opening initiated successfully!"))
-                (info_card(
-                    "Channel Details",
-                    vec![
-                        ("Temporary Channel ID", user_channel_id.0.to_string()),
-                        ("Node ID", form.node_id),
-                        ("Amount", format_sats_as_btc(form.amount_sats)),
-                        ("Push Amount", form.push_btc.map(format_sats_as_btc).unwrap_or_else(|| "₿ 0".to_string())),
-                    ]
-                ))
-                div class="card" {
-                    p { "The channel is now being opened. It may take some time for the channel to become active." }
-                    a href="/balance" { button { "← Back to Lightning" } }
-                }
-            }
-        }
-        Err(e) => {
-            tracing::error!("Web interface: Failed to open channel to {}: {}", pubkey, e);
-            html! {
-                (error_message(&format!("Failed to open channel: {e}")))
-                div class="card" {
-                    a href="/channels/open" { button { "← Try Again" } }
-                }
-            }
-        }
-    };
-
-    Ok(Response::builder()
-        .header("content-type", "text/html")
-        .body(Body::from(
-            layout("Open Channel Result", content).into_string(),
-        ))
-        .unwrap())
-}
-
-pub async fn close_channel_page(
-    State(_state): State<AppState>,
-    query: Query<HashMap<String, String>>,
-) -> Result<Html<String>, StatusCode> {
-    let channel_id = query.get("channel_id").unwrap_or(&"".to_string()).clone();
-    let node_id = query.get("node_id").unwrap_or(&"".to_string()).clone();
-
-    if channel_id.is_empty() || node_id.is_empty() {
-        let content = html! {
-            (error_message("Missing channel ID or node ID"))
-            div class="card" {
-                a href="/balance" { button { "← Back to Lightning" } }
-            }
-        };
-        return Ok(Html(layout("Close Channel Error", content).into_string()));
-    }
-
-    let content = form_card(
-        "Close Channel",
-        html! {
-            p { "Are you sure you want to close this channel?" }
-            div class="info-item" {
-                span class="info-label" { "User Channel ID:" }
-                span class="info-value" style="font-family: monospace; font-size: 0.85rem;" { (channel_id) }
-            }
-            div class="info-item" {
-                span class="info-label" { "Node ID:" }
-                span class="info-value" style="font-family: monospace; font-size: 0.85rem;" { (node_id) }
-            }
-            form method="post" action="/channels/close" style="margin-top: 1rem;" {
-                input type="hidden" name="channel_id" value=(channel_id) {}
-                input type="hidden" name="node_id" value=(node_id) {}
-                button type="submit" style="background: #dc3545;" { "Close Channel" }
-                " "
-                a href="/balance" { button type="button" { "Cancel" } }
-            }
-        },
-    );
-
-    Ok(Html(layout("Close Channel", content).into_string()))
-}
-
-pub async fn force_close_channel_page(
-    State(_state): State<AppState>,
-    query: Query<HashMap<String, String>>,
-) -> Result<Html<String>, StatusCode> {
-    let channel_id = query.get("channel_id").unwrap_or(&"".to_string()).clone();
-    let node_id = query.get("node_id").unwrap_or(&"".to_string()).clone();
-
-    if channel_id.is_empty() || node_id.is_empty() {
-        let content = html! {
-            (error_message("Missing channel ID or node ID"))
-            div class="card" {
-                a href="/balance" { button { "← Back to Lightning" } }
-            }
-        };
-        return Ok(Html(
-            layout("Force Close Channel Error", content).into_string(),
-        ));
-    }
-
-    let content = form_card(
-        "Force Close Channel",
-        html! {
-            div style="border: 2px solid #d63384; background-color: rgba(214, 51, 132, 0.1); padding: 1rem; margin-bottom: 1rem; border-radius: 0.5rem;" {
-                h4 style="color: #d63384; margin: 0 0 0.5rem 0;" { "⚠️ Warning: Force Close" }
-                p style="color: #d63384; margin: 0; font-size: 0.9rem;" {
-                    "Force close should NOT be used if normal close is preferred. "
-                    "Force close will immediately broadcast the latest commitment transaction and may result in delayed fund recovery. "
-                    "Only use this if the channel counterparty is unresponsive or there are other issues preventing normal closure."
-                }
-            }
-            p { "Are you sure you want to force close this channel?" }
-            div class="info-item" {
-                span class="info-label" { "User Channel ID:" }
-                span class="info-value" style="font-family: monospace; font-size: 0.85rem;" { (channel_id) }
-            }
-            div class="info-item" {
-                span class="info-label" { "Node ID:" }
-                span class="info-value" style="font-family: monospace; font-size: 0.85rem;" { (node_id) }
-            }
-            form method="post" action="/channels/force-close" style="margin-top: 1rem;" {
-                input type="hidden" name="channel_id" value=(channel_id) {}
-                input type="hidden" name="node_id" value=(node_id) {}
-                button type="submit" style="background: #d63384;" { "Force Close Channel" }
-                " "
-                a href="/balance" { button type="button" { "Cancel" } }
-            }
-        },
-    );
-
-    Ok(Html(layout("Force Close Channel", content).into_string()))
-}
-
-#[derive(Deserialize)]
-pub struct CloseChannelForm {
-    channel_id: String,
-    node_id: String,
-}
-
-pub async fn post_close_channel(
-    State(state): State<AppState>,
-    Form(form): Form<CloseChannelForm>,
-) -> Result<Response, StatusCode> {
-    tracing::info!(
-        "Web interface: Attempting to close channel_id={} with node_id={}",
-        form.channel_id,
-        form.node_id
-    );
-
-    let node_pubkey = match PublicKey::from_str(&form.node_id) {
-        Ok(pk) => pk,
-        Err(e) => {
-            tracing::warn!(
-                "Web interface: Invalid node public key for channel close: {}",
-                e
-            );
-            let content = html! {
-                (error_message(&format!("Invalid node public key: {e}")))
-                div class="card" {
-                    a href="/channels" { button { "← Back to Channels" } }
-                }
-            };
-            return Ok(Response::builder()
-                .status(StatusCode::BAD_REQUEST)
-                .header("content-type", "text/html")
-                .body(Body::from(
-                    layout("Close Channel Error", content).into_string(),
-                ))
-                .unwrap());
-        }
-    };
-
-    let channel_id: u128 = match form.channel_id.parse() {
-        Ok(id) => id,
-        Err(e) => {
-            tracing::warn!("Web interface: Invalid channel ID for channel close: {}", e);
-            let content = html! {
-                (error_message(&format!("Invalid channel ID: {e}")))
-                div class="card" {
-                    a href="/channels" { button { "← Back to Channels" } }
-                }
-            };
-            return Ok(Response::builder()
-                .status(StatusCode::BAD_REQUEST)
-                .header("content-type", "text/html")
-                .body(Body::from(
-                    layout("Close Channel Error", content).into_string(),
-                ))
-                .unwrap());
-        }
-    };
-
-    let user_channel_id = UserChannelId(channel_id);
-    tracing::info!(
-        "Web interface: Initiating cooperative close for channel {} with {}",
-        channel_id,
-        node_pubkey
-    );
-    let close_result = state
-        .node
-        .inner
-        .close_channel(&user_channel_id, node_pubkey);
-
-    let content = match close_result {
-        Ok(()) => {
-            tracing::info!(
-                "Web interface: Successfully initiated cooperative close for channel {} with {}",
-                channel_id,
-                node_pubkey
-            );
-            html! {
-                (success_message("Channel closing initiated successfully!"))
-                div class="card" {
-                    p { "The channel is now being closed. It may take some time for the closing transaction to be confirmed." }
-                    a href="/balance" { button { "← Back to Lightning" } }
-                }
-            }
-        }
-        Err(e) => {
-            tracing::error!(
-                "Web interface: Failed to close channel {} with {}: {}",
-                channel_id,
-                node_pubkey,
-                e
-            );
-            html! {
-                (error_message(&format!("Failed to close channel: {e}")))
-                div class="card" {
-                    a href="/balance" { button { "← Back to Lightning" } }
-                }
-            }
-        }
-    };
-
-    Ok(Response::builder()
-        .header("content-type", "text/html")
-        .body(Body::from(
-            layout("Close Channel Result", content).into_string(),
-        ))
-        .unwrap())
-}
-
-pub async fn post_force_close_channel(
-    State(state): State<AppState>,
-    Form(form): Form<CloseChannelForm>,
-) -> Result<Response, StatusCode> {
-    tracing::info!(
-        "Web interface: Attempting to FORCE CLOSE channel_id={} with node_id={}",
-        form.channel_id,
-        form.node_id
-    );
-
-    let node_pubkey = match PublicKey::from_str(&form.node_id) {
-        Ok(pk) => pk,
-        Err(e) => {
-            tracing::warn!(
-                "Web interface: Invalid node public key for force close: {}",
-                e
-            );
-            let content = html! {
-                (error_message(&format!("Invalid node public key: {e}")))
-                div class="card" {
-                    a href="/channels" { button { "← Back to Channels" } }
-                }
-            };
-            return Ok(Response::builder()
-                .status(StatusCode::BAD_REQUEST)
-                .header("content-type", "text/html")
-                .body(Body::from(
-                    layout("Force Close Channel Error", content).into_string(),
-                ))
-                .unwrap());
-        }
-    };
-
-    let channel_id: u128 = match form.channel_id.parse() {
-        Ok(id) => id,
-        Err(e) => {
-            tracing::warn!("Web interface: Invalid channel ID for force close: {}", e);
-            let content = html! {
-                (error_message(&format!("Invalid channel ID: {e}")))
-                div class="card" {
-                    a href="/channels" { button { "← Back to Channels" } }
-                }
-            };
-            return Ok(Response::builder()
-                .status(StatusCode::BAD_REQUEST)
-                .header("content-type", "text/html")
-                .body(Body::from(
-                    layout("Force Close Channel Error", content).into_string(),
-                ))
-                .unwrap());
-        }
-    };
-
-    let user_channel_id = UserChannelId(channel_id);
-    tracing::warn!("Web interface: Initiating FORCE CLOSE for channel {} with {} - this will broadcast the latest commitment transaction", channel_id, node_pubkey);
-    let force_close_result =
-        state
-            .node
-            .inner
-            .force_close_channel(&user_channel_id, node_pubkey, None);
-
-    let content = match force_close_result {
-        Ok(()) => {
-            tracing::info!(
-                "Web interface: Successfully initiated force close for channel {} with {}",
-                channel_id,
-                node_pubkey
-            );
-            html! {
-                (success_message("Channel force close initiated successfully!"))
-                div class="card" style="border: 1px solid #d63384; background-color: rgba(214, 51, 132, 0.1);" {
-                    h4 style="color: #d63384;" { "Force Close Complete" }
-                    p { "The channel has been force closed. The latest commitment transaction has been broadcast to the network." }
-                    p style="color: #d63384; font-size: 0.9rem;" {
-                        "Note: Your funds may be subject to a time delay before they can be spent. "
-                        "This delay depends on the channel configuration and may be several blocks."
-                    }
-                    a href="/balance" { button { "← Back to Lightning" } }
-                }
-            }
-        }
-        Err(e) => {
-            tracing::error!(
-                "Web interface: Failed to force close channel {} with {}: {}",
-                channel_id,
-                node_pubkey,
-                e
-            );
-            html! {
-                (error_message(&format!("Failed to force close channel: {e}")))
-                div class="card" {
-                    a href="/balance" { button { "← Back to Lightning" } }
-                }
-            }
-        }
-    };
-
-    Ok(Response::builder()
-        .header("content-type", "text/html")
-        .body(Body::from(
-            layout("Force Close Channel Result", content).into_string(),
-        ))
-        .unwrap())
-}
-
-pub async fn invoices_page(State(_state): State<AppState>) -> Result<Html<String>, StatusCode> {
-    let content = html! {
-        h2 style="text-align: center; margin-bottom: 3rem;" { "Invoices" }
-        div class="grid" {
-            (form_card(
-                "Create BOLT11 Invoice",
-                html! {
-                    form method="post" action="/invoices/bolt11" {
-                        div class="form-group" {
-                            label for="amount_btc" { "Amount" }
-                            input type="number" id="amount_btc" name="amount_btc" required placeholder="₿0" step="0.00000001" {}
-                        }
-                        div class="form-group" {
-                            label for="description" { "Description (optional)" }
-                            input type="text" id="description" name="description" placeholder="Payment for..." {}
-                        }
-                        div class="form-group" {
-                            label for="expiry_seconds" { "Expiry (seconds, optional)" }
-                            input type="number" id="expiry_seconds" name="expiry_seconds" placeholder="3600" {}
-                        }
-                        button type="submit" { "Create BOLT11 Invoice" }
-                    }
-                }
-            ))
-
-            (form_card(
-                "Create BOLT12 Offer",
-                html! {
-                    form method="post" action="/invoices/bolt12" {
-                        div class="form-group" {
-                            label for="amount_btc" { "Amount (optional for variable amount)" }
-                            input type="number" id="amount_btc" name="amount_btc" placeholder="₿0" step="0.00000001" {}
-                        }
-                        div class="form-group" {
-                            label for="description" { "Description (optional)" }
-                            input type="text" id="description" name="description" placeholder="Payment for..." {}
-                        }
-                        div class="form-group" {
-                            label for="expiry_seconds" { "Expiry (seconds, optional)" }
-                            input type="number" id="expiry_seconds" name="expiry_seconds" placeholder="3600" {}
-                        }
-                        button type="submit" { "Create BOLT12 Offer" }
-                    }
-                }
-            ))
-        }
-    };
-
-    Ok(Html(layout("Create Invoices", content).into_string()))
-}
-
-#[derive(Deserialize)]
-pub struct CreateBolt11Form {
-    amount_btc: u64,
-    description: Option<String>,
-    #[serde(deserialize_with = "deserialize_optional_u32")]
-    expiry_seconds: Option<u32>,
-}
-
-pub async fn post_create_bolt11(
-    State(state): State<AppState>,
-    Form(form): Form<CreateBolt11Form>,
-) -> Result<Response, StatusCode> {
-    use ldk_node::lightning_invoice::{Bolt11InvoiceDescription, Description};
-
-    tracing::info!(
-        "Web interface: Creating BOLT11 invoice for amount={} sats, description={:?}, expiry={}s",
-        form.amount_btc,
-        form.description,
-        form.expiry_seconds.unwrap_or(3600)
-    );
-
-    // Handle optional description
-    let description_text = form.description.clone().unwrap_or_else(|| "".to_string());
-    let description = if description_text.is_empty() {
-        // Use empty description for empty or missing description
-        match Description::new("".to_string()) {
-            Ok(desc) => Bolt11InvoiceDescription::Direct(desc),
-            Err(_) => {
-                // Fallback to a minimal valid description
-                let desc = Description::new(" ".to_string()).unwrap();
-                Bolt11InvoiceDescription::Direct(desc)
-            }
-        }
-    } else {
-        match Description::new(description_text.clone()) {
-            Ok(desc) => Bolt11InvoiceDescription::Direct(desc),
-            Err(e) => {
-                tracing::warn!(
-                    "Web interface: Invalid description for BOLT11 invoice: {}",
-                    e
-                );
-                let content = html! {
-                    (error_message(&format!("Invalid description: {e}")))
-                    div class="card" {
-                        a href="/invoices" { button { "← Try Again" } }
-                    }
-                };
-                return Ok(Response::builder()
-                    .status(StatusCode::BAD_REQUEST)
-                    .header("content-type", "text/html")
-                    .body(Body::from(
-                        layout("Create Invoice Error", content).into_string(),
-                    ))
-                    .unwrap());
-            }
-        }
-    };
-
-    // Convert Bitcoin to millisatoshis
-    let amount_msats = form.amount_btc * 1_000;
-
-    let expiry_seconds = form.expiry_seconds.unwrap_or(3600);
-    let invoice_result =
-        state
-            .node
-            .inner
-            .bolt11_payment()
-            .receive(amount_msats, &description, expiry_seconds);
-
-    let content = match invoice_result {
-        Ok(invoice) => {
-            tracing::info!(
-                "Web interface: Successfully created BOLT11 invoice with payment_hash={}",
-                invoice.payment_hash()
-            );
-            let current_time = std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .unwrap_or_default()
-                .as_secs();
-
-            let description_display = if description_text.is_empty() {
-                "None".to_string()
-            } else {
-                description_text.clone()
-            };
-
-            html! {
-                (success_message("BOLT11 Invoice created successfully!"))
-                (info_card(
-                    "Invoice Details",
-                    vec![
-                        ("Payment Hash", invoice.payment_hash().to_string()),
-                        ("Amount", format_sats_as_btc(form.amount_btc)),
-                        ("Description", description_display),
-                        ("Expires At", format!("{}", current_time + expiry_seconds as u64)),
-                    ]
-                ))
-                div class="card" {
-                    h3 { "Invoice (copy this to share)" }
-                    textarea readonly style="width: 100%; height: 150px; font-family: monospace; font-size: 0.8rem;" {
-                        (invoice.to_string())
-                    }
-                }
-                div class="card" {
-                    a href="/invoices" { button { "← Create Another Invoice" } }
-                }
-            }
-        }
-        Err(e) => {
-            tracing::error!("Web interface: Failed to create BOLT11 invoice: {}", e);
-            html! {
-                (error_message(&format!("Failed to create invoice: {e}")))
-                div class="card" {
-                    a href="/invoices" { button { "← Try Again" } }
-                }
-            }
-        }
-    };
-
-    Ok(Response::builder()
-        .header("content-type", "text/html")
-        .body(Body::from(
-            layout("BOLT11 Invoice Created", content).into_string(),
-        ))
-        .unwrap())
-}
-
-#[derive(Deserialize)]
-pub struct CreateBolt12Form {
-    #[serde(deserialize_with = "deserialize_optional_f64")]
-    amount_btc: Option<f64>,
-    description: Option<String>,
-    #[serde(deserialize_with = "deserialize_optional_u32")]
-    expiry_seconds: Option<u32>,
-}
-
-pub async fn post_create_bolt12(
-    State(state): State<AppState>,
-    Form(form): Form<CreateBolt12Form>,
-) -> Result<Response, StatusCode> {
-    let expiry_seconds = form.expiry_seconds.unwrap_or(3600);
-    let description_text = form.description.unwrap_or_else(|| "".to_string());
-
-    tracing::info!(
-        "Web interface: Creating BOLT12 offer for amount={:?} btc, description={:?}, expiry={}s",
-        form.amount_btc,
-        description_text,
-        expiry_seconds
-    );
-
-    let offer_result = if let Some(amount_btc) = form.amount_btc {
-        // Convert Bitcoin to millisatoshis (1 BTC = 100,000,000,000 msats)
-        let amount_msats = (amount_btc * 100_000_000_000.0) as u64;
-        state.node.inner.bolt12_payment().receive(
-            amount_msats,
-            &description_text,
-            Some(expiry_seconds),
-            None,
-        )
-    } else {
-        state
-            .node
-            .inner
-            .bolt12_payment()
-            .receive_variable_amount(&description_text, Some(expiry_seconds))
-    };
-
-    let content = match offer_result {
-        Ok(offer) => {
-            tracing::info!(
-                "Web interface: Successfully created BOLT12 offer with offer_id={}",
-                offer.id()
-            );
-            let current_time = std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .unwrap_or_default()
-                .as_secs();
-
-            let amount_display = form
-                .amount_btc
-                .map(|a| format_sats_as_btc((a * 100_000_000.0) as u64))
-                .unwrap_or_else(|| "Variable amount".to_string());
-
-            let description_display = if description_text.is_empty() {
-                "None".to_string()
-            } else {
-                description_text
-            };
-
-            html! {
-                (success_message("BOLT12 Offer created successfully!"))
-                (info_card(
-                    "Offer Details",
-                    vec![
-                        ("Offer ID", offer.id().to_string()),
-                        ("Amount", amount_display),
-                        ("Description", description_display),
-                        ("Expires At", format!("{}", current_time + expiry_seconds as u64)),
-                    ]
-                ))
-                div class="card" {
-                    h3 { "Offer (copy this to share)" }
-                    textarea readonly style="width: 100%; height: 150px; font-family: monospace; font-size: 0.8rem;" {
-                        (offer.to_string())
-                    }
-                }
-                div class="card" {
-                    a href="/invoices" { button { "← Create Another Offer" } }
-                }
-            }
-        }
-        Err(e) => {
-            tracing::error!("Web interface: Failed to create BOLT12 offer: {}", e);
-            html! {
-                (error_message(&format!("Failed to create offer: {e}")))
-                div class="card" {
-                    a href="/invoices" { button { "← Try Again" } }
-                }
-            }
-        }
-    };
-
-    Ok(Response::builder()
-        .header("content-type", "text/html")
-        .body(Body::from(
-            layout("BOLT12 Offer Created", content).into_string(),
-        ))
-        .unwrap())
-}
-
-#[derive(Deserialize)]
-pub struct PaymentsQuery {
-    filter: Option<String>,
-    page: Option<u32>,
-    per_page: Option<u32>,
-}
-
-pub async fn payments_page(
-    State(state): State<AppState>,
-    query: Query<PaymentsQuery>,
-) -> Result<Html<String>, StatusCode> {
-    let filter = query.filter.as_deref().unwrap_or("all");
-    let page = query.page.unwrap_or(1).max(1);
-    let per_page = query.per_page.unwrap_or(25).clamp(10, 100); // Limit between 10-100 items per page
-
-    // Use efficient pagination function
-    let (current_page_payments, total_count) = get_paginated_payments_streaming(
-        &state.node.inner,
-        filter,
-        ((page - 1) * per_page) as usize,
-        per_page as usize,
-    );
-
-    // Calculate pagination
-    let total_pages = ((total_count as f64) / (per_page as f64)).ceil() as u32;
-    let start_index = ((page - 1) * per_page) as usize;
-    let end_index = (start_index + per_page as usize).min(total_count);
-
-    // Helper function to build URL with pagination params
-    let build_url = |new_page: u32, new_filter: &str, new_per_page: u32| -> String {
-        let mut params = vec![];
-        if new_filter != "all" {
-            params.push(format!("filter={}", new_filter));
-        }
-        if new_page != 1 {
-            params.push(format!("page={}", new_page));
-        }
-        if new_per_page != 25 {
-            params.push(format!("per_page={}", new_per_page));
-        }
-
-        if params.is_empty() {
-            "/payments".to_string()
-        } else {
-            format!("/payments?{}", params.join("&"))
-        }
-    };
-
-    let content = html! {
-        h2 style="text-align: center; margin-bottom: 3rem;" { "Payments" }
-        div class="card" {
-            div class="payment-list-header" {
-                div {
-                    h2 { "Payment History" }
-                    @if total_count > 0 {
-                        p style="margin: 0.25rem 0 0 0; color: #666; font-size: 0.9rem;" {
-                            "Showing " (start_index + 1) " to " (end_index) " of " (total_count) " payments"
-                        }
-                    }
-                }
-                div class="payment-filter-tabs" {
-                    a href=(build_url(1, "all", per_page)) class=(if filter == "all" { "payment-filter-tab active" } else { "payment-filter-tab" }) { "All" }
-                    a href=(build_url(1, "incoming", per_page)) class=(if filter == "incoming" { "payment-filter-tab active" } else { "payment-filter-tab" }) { "Incoming" }
-                    a href=(build_url(1, "outgoing", per_page)) class=(if filter == "outgoing" { "payment-filter-tab active" } else { "payment-filter-tab" }) { "Outgoing" }
-                }
-            }
-
-            // Payment list (no metrics here)
-            @if current_page_payments.is_empty() {
-                @if total_count == 0 {
-                    p { "No payments found." }
-                } @else {
-                    p { "No payments found on this page. "
-                        a href=(build_url(1, filter, per_page)) { "Go to first page" }
-                    }
-                }
-            } @else {
-                @for payment in &current_page_payments {
-                    @let direction_str = match payment.direction {
-                        PaymentDirection::Inbound => "Inbound",
-                        PaymentDirection::Outbound => "Outbound",
-                    };
-
-                    @let status_str = match payment.status {
-                        PaymentStatus::Pending => "Pending",
-                        PaymentStatus::Succeeded => "Succeeded",
-                        PaymentStatus::Failed => "Failed",
-                    };
-
-                    @let amount_str = payment.amount_msat.map(format_msats_as_btc).unwrap_or_else(|| "Unknown".to_string());
-
-                    @let (payment_hash, description, payment_type, preimage) = match &payment.kind {
-                        PaymentKind::Bolt11 { hash, preimage, .. } => {
-                            (Some(hash.to_string()), None::<String>, "BOLT11", preimage.map(|p| p.to_string()))
-                        },
-                        PaymentKind::Bolt12Offer { hash, offer_id, preimage, .. } => {
-                            // For BOLT12, we can use either the payment hash or offer ID
-                            let identifier = hash.map(|h| h.to_string()).unwrap_or_else(|| offer_id.to_string());
-                            (Some(identifier), None::<String>, "BOLT12", preimage.map(|p| p.to_string()))
-                        },
-                        PaymentKind::Bolt12Refund { hash, preimage, .. } => {
-                            (hash.map(|h| h.to_string()), None::<String>, "BOLT12", preimage.map(|p| p.to_string()))
-                        },
-                        PaymentKind::Spontaneous { hash, preimage, .. } => {
-                            (Some(hash.to_string()), None::<String>, "Spontaneous", preimage.map(|p| p.to_string()))
-                        },
-                        PaymentKind::Onchain { txid, .. } => {
-                            (Some(txid.to_string()), None::<String>, "On-chain", None)
-                        },
-                        PaymentKind::Bolt11Jit { hash, .. } => {
-                            (Some(hash.to_string()), None::<String>, "BOLT11 JIT", None)
-                        },
-                    };
-
-                    (payment_list_item(
-                        &payment.id.to_string(),
-                        direction_str,
-                        status_str,
-                        &amount_str,
-                        payment_hash.as_deref(),
-                        description.as_deref(),
-                        Some(payment.latest_update_timestamp), // Use the actual timestamp
-                        payment_type,
-                        preimage.as_deref(),
-                    ))
-                }
-            }
-
-            // Pagination controls (bottom)
-            @if total_pages > 1 {
-                div class="pagination-controls" style="margin-top: 2rem; padding-top: 1rem; border-top: 1px solid #eee;" {
-                    div class="pagination" style="display: flex; justify-content: center; align-items: center; gap: 0.5rem;" {
-                        // Previous page
-                        @if page > 1 {
-                            a href=(build_url(page - 1, filter, per_page)) class="pagination-btn" { "← Previous" }
-                        } @else {
-                            span class="pagination-btn disabled" { "← Previous" }
-                        }
-
-                        // Page numbers
-                        @let start_page = (page.saturating_sub(2)).max(1);
-                        @let end_page = (page + 2).min(total_pages);
-
-                        @if start_page > 1 {
-                            a href=(build_url(1, filter, per_page)) class="pagination-number" { "1" }
-                            @if start_page > 2 {
-                                span class="pagination-ellipsis" { "..." }
-                            }
-                        }
-
-                        @for p in start_page..=end_page {
-                            @if p == page {
-                                span class="pagination-number active" { (p) }
-                            } @else {
-                                a href=(build_url(p, filter, per_page)) class="pagination-number" { (p) }
-                            }
-                        }
-
-                        @if end_page < total_pages {
-                            @if end_page < total_pages - 1 {
-                                span class="pagination-ellipsis" { "..." }
-                            }
-                            a href=(build_url(total_pages, filter, per_page)) class="pagination-number" { (total_pages) }
-                        }
-
-                        // Next page
-                        @if page < total_pages {
-                            a href=(build_url(page + 1, filter, per_page)) class="pagination-btn" { "Next →" }
-                        } @else {
-                            span class="pagination-btn disabled" { "Next →" }
-                        }
-                    }
-                }
-            }
-
-            // Compact per-page selector integrated with pagination
-            @if total_count > 0 {
-                div class="per-page-selector" {
-                    label for="per-page" { "Show:" }
-                    select id="per-page" onchange="changePage()" {
-                        option value="10" selected[per_page == 10] { "10" }
-                        option value="25" selected[per_page == 25] { "25" }
-                        option value="50" selected[per_page == 50] { "50" }
-                        option value="100" selected[per_page == 100] { "100" }
-                    }
-                    span { "per page" }
-                }
-            }
-        }
-
-        // JavaScript for per-page selector
-        script {
-            "function changePage() {
-                const perPageSelect = document.getElementById('per-page');
-                const newPerPage = perPageSelect.value;
-                const currentUrl = new URL(window.location);
-                currentUrl.searchParams.set('per_page', newPerPage);
-                currentUrl.searchParams.set('page', '1'); // Reset to first page when changing per_page
-                window.location.href = currentUrl.toString();
-            }"
-        }
-    };
-
-    Ok(Html(layout("Payment History", content).into_string()))
-}
-
-pub async fn send_payments_page(
-    State(_state): State<AppState>,
-) -> Result<Html<String>, StatusCode> {
-    let content = html! {
-        h2 style="text-align: center; margin-bottom: 3rem;" { "Send Payment" }
-        div class="grid" {
-            (form_card(
-                "Pay BOLT11 Invoice",
-                html! {
-                    form method="post" action="/payments/bolt11" {
-                        div class="form-group" {
-                            label for="invoice" { "BOLT11 Invoice" }
-                            textarea id="invoice" name="invoice" required placeholder="lnbc..." style="height: 120px;" {}
-                        }
-                        div class="form-group" {
-                            label for="amount_btc" { "Amount Override (optional)" }
-                            input type="number" id="amount_btc" name="amount_btc" placeholder="Leave empty to use invoice amount" step="1" {}
-                        }
-                        button type="submit" { "Pay BOLT11 Invoice" }
-                    }
-                }
-            ))
-
-            (form_card(
-                "Pay BOLT12 Offer",
-                html! {
-                    form method="post" action="/payments/bolt12" {
-                        div class="form-group" {
-                            label for="offer" { "BOLT12 Offer" }
-                            textarea id="offer" name="offer" required placeholder="lno..." style="height: 120px;" {}
-                        }
-                        div class="form-group" {
-                            label for="amount_btc" { "Amount (required for variable amount offers)" }
-                            input type="number" id="amount_btc" name="amount_btc" placeholder="Required for variable amount offers, ignored for fixed amount offers" step="1" {}
-                        }
-                        button type="submit" { "Pay BOLT12 Offer" }
-                    }
-                }
-            ))
-        }
-
-        div class="card" {
-            h3 { "Payment History" }
-            a href="/payments" { button { "View All Payments" } }
-        }
-    };
-
-    Ok(Html(layout("Send Payments", content).into_string()))
-}
-
-#[derive(Debug, Deserialize)]
-pub struct PayBolt11Form {
-    invoice: String,
-    #[serde(deserialize_with = "deserialize_optional_u64")]
-    amount_btc: Option<u64>,
-}
-
-pub async fn post_pay_bolt11(
-    State(state): State<AppState>,
-    Form(form): Form<PayBolt11Form>,
-) -> Result<Response, StatusCode> {
-    let invoice = match Bolt11Invoice::from_str(form.invoice.trim()) {
-        Ok(inv) => inv,
-        Err(e) => {
-            tracing::warn!("Web interface: Invalid BOLT11 invoice provided: {}", e);
-            let content = html! {
-                (error_message(&format!("Invalid BOLT11 invoice: {e}")))
-                div class="card" {
-                    a href="/payments" { button { "← Try Again" } }
-                }
-            };
-            return Ok(Response::builder()
-                .status(StatusCode::BAD_REQUEST)
-                .header("content-type", "text/html")
-                .body(Body::from(layout("Payment Error", content).into_string()))
-                .unwrap());
-        }
-    };
-
-    tracing::info!(
-        "Web interface: Attempting to pay BOLT11 invoice payment_hash={}, amount_override={:?}",
-        invoice.payment_hash(),
-        form.amount_btc
-    );
-
-    let payment_id = if let Some(amount_btc) = form.amount_btc {
-        // Convert Bitcoin to millisatoshis
-        let amount_msats = amount_btc * 1000;
-        state
-            .node
-            .inner
-            .bolt11_payment()
-            .send_using_amount(&invoice, amount_msats, None)
-    } else {
-        state.node.inner.bolt11_payment().send(&invoice, None)
-    };
-
-    let payment_id = match payment_id {
-        Ok(id) => {
-            tracing::info!(
-                "Web interface: BOLT11 payment initiated with payment_id={}",
-                hex::encode(id.0)
-            );
-            id
-        }
-        Err(e) => {
-            tracing::error!("Web interface: Failed to initiate BOLT11 payment: {}", e);
-            let content = html! {
-                (error_message(&format!("Failed to initiate payment: {e}")))
-                div class="card" {
-                    a href="/payments" { button { "← Try Again" } }
-                }
-            };
-            return Ok(Response::builder()
-                .status(StatusCode::INTERNAL_SERVER_ERROR)
-                .header("content-type", "text/html")
-                .body(Body::from(layout("Payment Error", content).into_string()))
-                .unwrap());
-        }
-    };
-
-    // Wait for payment to complete (max 10 seconds)
-    let start = std::time::Instant::now();
-    let timeout = std::time::Duration::from_secs(10);
-
-    let payment_result = loop {
-        if let Some(details) = state.node.inner.payment(&payment_id) {
-            match details.status {
-                PaymentStatus::Succeeded => {
-                    tracing::info!(
-                        "Web interface: BOLT11 payment succeeded for payment_hash={}",
-                        invoice.payment_hash()
-                    );
-                    break Ok(details);
-                }
-                PaymentStatus::Failed => {
-                    tracing::error!(
-                        "Web interface: BOLT11 payment failed for payment_hash={}",
-                        invoice.payment_hash()
-                    );
-                    break Err("Payment failed".to_string());
-                }
-                PaymentStatus::Pending => {
-                    if start.elapsed() > timeout {
-                        tracing::warn!(
-                            "Web interface: BOLT11 payment timeout for payment_hash={}",
-                            invoice.payment_hash()
-                        );
-                        break Err("Payment is still pending after timeout".to_string());
-                    }
-                    tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-                    continue;
-                }
-            }
-        } else {
-            break Err("Payment not found".to_string());
-        }
-    };
-
-    let content = match payment_result {
-        Ok(details) => {
-            let (preimage, fee_msats) = match details.kind {
-                PaymentKind::Bolt11 {
-                    hash: _,
-                    preimage,
-                    secret: _,
-                } => (
-                    preimage.map(|p| p.to_string()).unwrap_or_default(),
-                    details.fee_paid_msat.unwrap_or(0),
-                ),
-                _ => (String::new(), 0),
-            };
-
-            html! {
-                (success_message("Payment succeeded!"))
-                (info_card(
-                    "Payment Details",
-                    vec![
-                        ("Payment Hash", invoice.payment_hash().to_string()),
-                        ("Payment Preimage", preimage),
-                        ("Fee Paid", format_msats_as_btc(fee_msats)),
-                        ("Amount", form.amount_btc.map(|_a| format_sats_as_btc(details.amount_msat.unwrap_or(1000) / 1000)).unwrap_or_default()),
-                    ]
-                ))
-                div class="card" {
-                    a href="/payments" { button { "← Make Another Payment" } }
-                }
-            }
-        }
-        Err(error) => {
-            html! {
-                (error_message(&format!("Payment failed: {error}")))
-                div class="card" {
-                    a href="/payments" { button { "← Try Again" } }
-                }
-            }
-        }
-    };
-
-    Ok(Response::builder()
-        .header("content-type", "text/html")
-        .body(Body::from(layout("Payment Result", content).into_string()))
-        .unwrap())
-}
-
-#[derive(Deserialize)]
-pub struct PayBolt12Form {
-    offer: String,
-    #[serde(deserialize_with = "deserialize_optional_u64")]
-    amount_btc: Option<u64>,
-}
-
-pub async fn post_pay_bolt12(
-    State(state): State<AppState>,
-    Form(form): Form<PayBolt12Form>,
-) -> Result<Response, StatusCode> {
-    let offer = match Offer::from_str(form.offer.trim()) {
-        Ok(offer) => offer,
-        Err(e) => {
-            tracing::warn!("Web interface: Invalid BOLT12 offer provided: {:?}", e);
-            let content = html! {
-                (error_message(&format!("Invalid BOLT12 offer: {e:?}")))
-                div class="card" {
-                    a href="/payments" { button { "← Try Again" } }
-                }
-            };
-            return Ok(Response::builder()
-                .status(StatusCode::BAD_REQUEST)
-                .header("content-type", "text/html")
-                .body(Body::from(layout("Payment Error", content).into_string()))
-                .unwrap());
-        }
-    };
-
-    tracing::info!(
-        "Web interface: Attempting to pay BOLT12 offer offer_id={}, amount_override={:?}",
-        offer.id(),
-        form.amount_btc
-    );
-
-    // Determine payment method based on offer type and user input
-    let payment_id = match offer.amount() {
-        Some(_) => {
-            // Fixed amount offer - use send() method, ignore user input amount
-            state.node.inner.bolt12_payment().send(&offer, None, None)
-        }
-        None => {
-            // Variable amount offer - requires user to specify amount via send_using_amount()
-            let amount_btc = match form.amount_btc {
-                Some(amount) => amount,
-                None => {
-                    tracing::warn!("Web interface: Amount required for variable amount BOLT12 offer but not provided");
-                    let content = html! {
-                        (error_message("Amount is required for variable amount offers. This offer does not have a fixed amount, so you must specify how much you want to pay."))
-                        div class="card" {
-                            a href="/payments" { button { "← Try Again" } }
-                        }
-                    };
-                    return Ok(Response::builder()
-                        .status(StatusCode::BAD_REQUEST)
-                        .header("content-type", "text/html")
-                        .body(Body::from(layout("Payment Error", content).into_string()))
-                        .unwrap());
-                }
-            };
-            let amount_msats = amount_btc * 1_000;
-            state
-                .node
-                .inner
-                .bolt12_payment()
-                .send_using_amount(&offer, amount_msats, None, None)
-        }
-    };
-
-    let payment_id = match payment_id {
-        Ok(id) => {
-            tracing::info!(
-                "Web interface: BOLT12 payment initiated with payment_id={}",
-                hex::encode(id.0)
-            );
-            id
-        }
-        Err(e) => {
-            tracing::error!("Web interface: Failed to initiate BOLT12 payment: {}", e);
-            let content = html! {
-                (error_message(&format!("Failed to initiate payment: {e}")))
-                div class="card" {
-                    a href="/payments" { button { "← Try Again" } }
-                }
-            };
-            return Ok(Response::builder()
-                .status(StatusCode::INTERNAL_SERVER_ERROR)
-                .header("content-type", "text/html")
-                .body(Body::from(layout("Payment Error", content).into_string()))
-                .unwrap());
-        }
-    };
-
-    // Wait for payment to complete (max 10 seconds)
-    let start = std::time::Instant::now();
-    let timeout = std::time::Duration::from_secs(10);
-
-    let payment_result = loop {
-        if let Some(details) = state.node.inner.payment(&payment_id) {
-            match details.status {
-                PaymentStatus::Succeeded => {
-                    tracing::info!(
-                        "Web interface: BOLT12 payment succeeded for offer_id={}",
-                        offer.id()
-                    );
-                    break Ok(details);
-                }
-                PaymentStatus::Failed => {
-                    tracing::error!(
-                        "Web interface: BOLT12 payment failed for offer_id={}",
-                        offer.id()
-                    );
-                    break Err("Payment failed".to_string());
-                }
-                PaymentStatus::Pending => {
-                    if start.elapsed() > timeout {
-                        tracing::warn!(
-                            "Web interface: BOLT12 payment timeout for offer_id={}",
-                            offer.id()
-                        );
-                        break Err("Payment is still pending after timeout".to_string());
-                    }
-                    tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-                    continue;
-                }
-            }
-        } else {
-            break Err("Payment not found".to_string());
-        }
-    };
-
-    let content = match payment_result {
-        Ok(details) => {
-            let (payment_hash, preimage, fee_msats) = match details.kind {
-                PaymentKind::Bolt12Offer {
-                    hash,
-                    preimage,
-                    secret: _,
-                    offer_id: _,
-                    payer_note: _,
-                    quantity: _,
-                } => (
-                    hash.map(|h| h.to_string()).unwrap_or_default(),
-                    preimage.map(|p| p.to_string()).unwrap_or_default(),
-                    details.fee_paid_msat.unwrap_or(0),
-                ),
-                _ => (String::new(), String::new(), 0),
-            };
-
-            html! {
-                (success_message("Payment succeeded!"))
-                (info_card(
-                    "Payment Details",
-                    vec![
-                        ("Payment Hash", payment_hash),
-                        ("Payment Preimage", preimage),
-                        ("Fee Paid", format_msats_as_btc(fee_msats)),
-                        ("Amount Paid", form.amount_btc.map(format_sats_as_btc).unwrap_or_else(|| {
-                            // If no amount was specified in the form, show the actual amount from the payment details
-                            details.amount_msat.map(format_msats_as_btc).unwrap_or_else(|| "Unknown".to_string())
-                        })),
-                    ]
-                ))
-                div class="card" {
-                    a href="/payments" { button { "← Make Another Payment" } }
-                }
-            }
-        }
-        Err(error) => {
-            html! {
-                (error_message(&format!("Payment failed: {error}")))
-                div class="card" {
-                    a href="/payments" { button { "← Try Again" } }
-                }
-            }
-        }
-    };
-
-    Ok(Response::builder()
-        .header("content-type", "text/html")
-        .body(Body::from(layout("Payment Result", content).into_string()))
-        .unwrap())
-}
