+++ crates/cdk-ldk-node/src/web/handlers/mod.rs
@@ -1,0 +1,17 @@
+pub mod channels;
+pub mod dashboard;
+pub mod invoices;
+pub mod lightning;
+pub mod onchain;
+pub mod payments;
+pub mod utils;
+
+// Re-export commonly used items
+// Re-export handler functions
+pub use channels::*;
+pub use dashboard::*;
+pub use invoices::*;
+pub use lightning::*;
+pub use onchain::*;
+pub use payments::*;
+pub use utils::AppState;
