+++ crates/cdk-ldk-node/src/web/templates/mod.rs
@@ -1,0 +1,10 @@
+pub mod components;
+pub mod formatters;
+pub mod layout;
+pub mod payments;
+
+// Re-export commonly used functions
+pub use components::*;
+pub use formatters::*;
+pub use layout::*;
+pub use payments::*;
